# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "高並行性イベントチケット予約システムのREST API",
        "title": "イベントチケット予約システム API",
        "termsOfService": "http://example.com/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/health": {
            "get": {
                "description": "アプリケーションの健全性を確認する",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "ヘルスチェック",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.HealthResponse"
                        }
                    }
                }
            }
        },
        "/events": {
            "get": {
                "description": "イベントの一覧を取得します",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "イベント一覧を取得",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "取得件数",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "オフセット",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handler.EventResponse"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "新しいイベントを作成します",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "イベントを作成",
                "parameters": [
                    {
                        "description": "イベント情報",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CreateEventRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handler.EventResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/events/{id}": {
            "get": {
                "description": "指定IDのイベントを取得します",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "イベントを取得",
                "parameters": [
                    {
                        "type": "string",
                        "description": "イベントID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.EventResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "指定IDのイベントを更新します",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "イベントを更新",
                "parameters": [
                    {
                        "type": "string",
                        "description": "イベントID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "イベント情報",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CreateEventRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.EventResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "指定IDのイベントを削除します",
                "tags": [
                    "events"
                ],
                "summary": "イベントを削除",
                "parameters": [
                    {
                        "type": "string",
                        "description": "イベントID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/reservations": {
            "get": {
                "description": "ログインユーザーの予約一覧を取得します",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reservations"
                ],
                "summary": "ユーザーの予約一覧を取得",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ユーザーID",
                        "name": "X-User-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "取得件数",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "オフセット",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handler.ReservationResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "座席を仮押さえします（15分間有効）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reservations"
                ],
                "summary": "予約を作成",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ユーザーID",
                        "name": "X-User-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "予約情報",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CreateReservationRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handler.ReservationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "座席が既に予約済み",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/reservations/{id}": {
            "get": {
                "description": "指定IDの予約を取得します",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reservations"
                ],
                "summary": "予約を取得",
                "parameters": [
                    {
                        "type": "string",
                        "description": "予約ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.ReservationResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/reservations/{id}/cancel": {
            "post": {
                "description": "予約をキャンセルし、座席を解放します",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reservations"
                ],
                "summary": "予約をキャンセル",
                "parameters": [
                    {
                        "type": "string",
                        "description": "予約ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.ReservationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/reservations/{id}/confirm": {
            "post": {
                "description": "仮押さえ中の予約を確定します",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reservations"
                ],
                "summary": "予約を確定",
                "parameters": [
                    {
                        "type": "string",
                        "description": "予約ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.ReservationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.CreateEventRequest": {
            "type": "object",
            "required": [
                "end_at",
                "name",
                "start_at",
                "total_seats"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "example": "年末スペシャルコンサート"
                },
                "end_at": {
                    "type": "string",
                    "example": "2025-12-31T21:00:00+09:00"
                },
                "name": {
                    "type": "string",
                    "example": "東京ドームコンサート2025"
                },
                "start_at": {
                    "type": "string",
                    "example": "2025-12-31T18:00:00+09:00"
                },
                "total_seats": {
                    "type": "integer",
                    "example": 50000
                },
                "venue": {
                    "type": "string",
                    "example": "東京ドーム"
                }
            }
        },
        "handler.CreateReservationRequest": {
            "type": "object",
            "required": [
                "event_id",
                "idempotency_key",
                "seat_ids"
            ],
            "properties": {
                "event_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "idempotency_key": {
                    "type": "string",
                    "example": "order-2025-001"
                },
                "seat_ids": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "seat-A1",
                        "seat-A2"
                    ]
                }
            }
        },
        "handler.EventResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2025-12-06T10:00:00+09:00"
                },
                "description": {
                    "type": "string",
                    "example": "年末スペシャルコンサート"
                },
                "end_at": {
                    "type": "string",
                    "example": "2025-12-31T21:00:00+09:00"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "name": {
                    "type": "string",
                    "example": "東京ドームコンサート2025"
                },
                "start_at": {
                    "type": "string",
                    "example": "2025-12-31T18:00:00+09:00"
                },
                "total_seats": {
                    "type": "integer",
                    "example": 50000
                },
                "updated_at": {
                    "type": "string",
                    "example": "2025-12-06T10:00:00+09:00"
                },
                "venue": {
                    "type": "string",
                    "example": "東京ドーム"
                }
            }
        },
        "handler.HealthResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "handler.ReservationResponse": {
            "type": "object",
            "properties": {
                "confirmed_at": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "event_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "expires_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "seat_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "seat-A1",
                        "seat-A2"
                    ]
                },
                "status": {
                    "type": "string",
                    "example": "pending"
                },
                "total_amount": {
                    "type": "integer",
                    "example": 10000
                },
                "user_id": {
                    "type": "string",
                    "example": "user-123"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-User-ID",
            "in": "header"
        }
    }
}